+++ merchstore/src/main/java/com/swift/merchstore/model/Product.java
@@ -19,7 +19,7 @@ public class Product {
     private Integer id;
     private @NotNull String name;
     private @NotNull String imageURL;
-    private @NotNull double price;
+    private @NotNull int price;
     private @NotNull String description;
 
     // Many-To-One relationship
@@ -51,11 +51,11 @@ public class Product {
         this.imageURL = imageURL;
     }
 
-    public double getPrice() {
+    public int getPrice() {
         return price;
     }
 
-    public void setPrice(double price) {
+    public void setPrice(int price) {
         this.price = price;
     }
